Added idempotence to most dl and configuration tasks
separated apt-get install for clarity

+++ officeonline-install.sh
@@ -36,34 +36,41 @@ dialog --backtitle "Information" \
 --msgbox 'THE INSTALLATION WILL TAKE REALLY VERY LONG TIME, 2-8 HOURS (It depends on the speed of your server), SO BE PATIENT PLEASE!!! You may see errors during the installation, just ignore them and let it do the work.' 10 78
 clear
 
-sed -i 's/# deb-src/deb-src/g' $soli
+grep -q '# deb-src' ${soli} && sed -i 's/# deb-src/deb-src/g' ${soli} && apt-get update
 
+apt-get install sudo curl libegl1-mesa-dev libkrb5-dev systemd python-polib git libkrb5-dev make openssl g++ libtool ccache libpng12-0 libpng12-dev libpcap0.8 libpcap0.8-dev libcunit1 libcunit1-dev libpng12-dev libcap-dev libtool m4 automake libcppunit-dev libcppunit-doc pkg-config wget libfontconfig1-dev -y
+[ $? -ne 0 ] && exit 1
+apt-get build-dep libreoffice -y
 
-apt-get update && apt-get upgrade -y
-
-apt-get install sudo curl libegl1-mesa-dev libkrb5-dev systemd python-polib git libkrb5-dev make openssl g++ libtool ccache libpng12-0 libpng12-dev libpcap0.8 libpcap0.8-dev libcunit1 libcunit1-dev libpng12-dev libcap-dev libtool m4 automake libcppunit-dev libcppunit-doc pkg-config wget libfontconfig1-dev  -y && sudo apt-get build-dep libreoffice -y
-
-curl -sL https://deb.nodesource.com/setup_6.x | bash -
-apt-get install nodejs -y
+if [ ! -f /etc/apt/sources.list.d/nodesource.list ]; then
+  curl -sL https://deb.nodesource.com/setup_6.x | bash -
+  apt-get install nodejs -y
+fi
 
 getent passwd lool || (useradd lool -G sudo; mkdir /home/lool)
 chown lool:lool /home/lool -R
 
-lo_version=$(curl -s http://download.documentfoundation.org/libreoffice/src/5.3.1/ | grep -oiE 'libreoffice-5.[0-9+]\.[0-9+]\.[0-9]' | awk 'NR == 1')
-wget -c http://download.documentfoundation.org/libreoffice/src/5.3.1/$lo_version.tar.xz -P /opt/
-tar xf /opt/$lo_version.tar.xz -C  /opt/
-mv /opt/$lo_version $ooo
+if [ ! -f ${ooo}/autogen.sh ]; then
+  lo_version=$(curl -s http://download.documentfoundation.org/libreoffice/src/5.3.0/ | grep -oiE 'libreoffice-5.[0-9+]\.[0-9+]\.[0-9]' | awk 'NR == 1')
+  [ ! -f $lo_version.tar.xz ] && wget -c http://download.documentfoundation.org/libreoffice/src/5.3.0/$lo_version.tar.xz -P /opt/
+  [ ! -d $lo_version ] && tar xf /opt/$lo_version.tar.xz -C  /opt/
+  mv /opt/$lo_version $ooo
+  chown lool:lool $ooo -R
+fi
 
-chown lool:lool $ooo -R
+###############################################################################
 
 sudo -H -u lool bash -c "for dir in ./ ; do (cd "$ooo" && $ooo/autogen.sh --without-help --without-myspell-dicts); done" | tee -a $log_file
 sudo -H -u lool bash -c "for dir in ./ ; do (cd "$ooo" && make); done" | tee -a $log_file
 
 poco_version=$(curl -s https://pocoproject.org/ | grep -oiE 'The latest stable release is [0-9+]\.[0-9\.]{1,}[0-9]{1,}' | awk '{print $NF}')
 poco="/opt/poco-${poco_version}-all"
-wget -c https://pocoproject.org/releases/poco-${poco_version}/poco-${poco_version}-all.tar.gz -P /opt/
-tar xf /opt/poco-${poco_version}-all.tar.gz -C  /opt/
-chown lool:lool $poco -R
+if [ ! -d $poco ]; then
+  [ ! -f /opt/poco-${poco_version}-all.tar.gz ] &&\
+  wget -c https://pocoproject.org/releases/poco-${poco_version}/poco-${poco_version}-all.tar.gz -P /opt/
+  tar xf /opt/poco-${poco_version}-all.tar.gz -C  /opt/
+  chown lool:lool $poco -R
+fi
 
 sudo -H -u lool bash -c "for dir in ./ ; do (cd "$poco" && ./configure); done" | tee -a $log_file
 sudo -H -u lool bash -c  "for dir in ./ ; do (cd "$poco" && make -j$cpu); done" | tee -a $log_file
@@ -94,7 +101,7 @@ PASSWORD=$(randpass 10 0)
 cat <<EOT > /lib/systemd/system/loolwsd.service
 
 [Unit]
-Description=LibreOffice On-Line WebSocket Daemon
+Description=LibreOffice OnLine WebSocket Daemon
 After=network.target
 
 [Service]
@@ -105,21 +112,24 @@ PermissionsStartOnly=true
 ExecStart=${oo}/loolwsd --o:sys_template_path=${oo}/systemplate --o:lo_template_path=${ooo}/instdir  --o:child_root_path=${oo}/jails --o:storage.filesystem[@allow]=true --o:admin_console.username=admin --o:admin_console.password=$PASSWORD
 User=lool
 KillMode=control-group
-Restart=always
+# Restart=always
 
 [Install]
 WantedBy=multi-user.target
 EOT
 
-mkdir /etc/loolwsd
-openssl genrsa -out /etc/loolwsd/key.pem 4096
-openssl req -out /etc/loolwsd/cert.csr -key /etc/loolwsd/key.pem -new -sha256 -nodes -subj "/C=DE/OU=onlineoffice-install.com/CN=onlineoffice-install.com/emailAddress=nomail@nodo.com"
-openssl x509 -req -days 365 -in /etc/loolwsd/cert.csr -signkey /etc/loolwsd/key.pem -out /etc/loolwsd/cert.pem
-openssl x509 -req -days 365 -in /etc/loolwsd/cert.csr -signkey /etc/loolwsd/key.pem -out /etc/loolwsd/ca-chain.cert.pem
-
-ln /lib/systemd/system/loolwsd.service /etc/systemd/system/loolwsd.service
-systemctl enable loolwsd.service
-
+if [ ! -f /etc/loolwsd/ca-chain.cert.pem ]; then
+  mkdir /etc/loolwsd
+  openssl genrsa -out /etc/loolwsd/key.pem 4096
+  openssl req -out /etc/loolwsd/cert.csr -key /etc/loolwsd/key.pem -new -sha256 -nodes -subj "/C=DE/OU=onlineoffice-install.com/CN=onlineoffice-install.com/emailAddress=nomail@nodo.com"
+  openssl x509 -req -days 365 -in /etc/loolwsd/cert.csr -signkey /etc/loolwsd/key.pem -out /etc/loolwsd/cert.pem
+  openssl x509 -req -days 365 -in /etc/loolwsd/cert.csr -signkey /etc/loolwsd/key.pem -out /etc/loolwsd/ca-chain.cert.pem
+fi
+if [! -e /etc/systemd/system/loolwsd.service ]; then
+  ln /lib/systemd/system/loolwsd.service /etc/systemd/system/loolwsd.service
+  systemctl enable loolwsd.service
+fi
+### Testing loolwsd ###
 dialog --backtitle "Information" \
 --title "Note" \
 --msgbox 'The installation log file is in '"${log_file}"'. After reboot you can use loolwsd.service using: systemctl (start,stop or status) loolwsd.service.
